+++ packages/bootstrap-4/src/IconButton/IconButton.tsx
@@ -1,8 +1,9 @@
 import React from "react";
 import { IconButtonProps } from "@rjsf/utils";
 import Button, { ButtonProps } from "react-bootstrap/Button";
-import { IoIosRemove } from "react-icons/io";
-import { AiOutlineArrowUp, AiOutlineArrowDown } from "react-icons/ai";
+import { IoIosRemove } from "@react-icons/all-files/io/IoIosRemove";
+import { AiOutlineArrowUp } from "@react-icons/all-files/ai/AiOutlineArrowUp";
+import { AiOutlineArrowDown } from "@react-icons/all-files/ai/AiOutlineArrowDown";
 
 const IconButton = (props: IconButtonProps & ButtonProps) => {
   const { icon, iconType, className, ...otherProps } = props;