+++ CHANGELOG.md
@@ -38,6 +38,9 @@ should change the heading of the (upcoming) version to include a major version b
 - Upgraded from the `1.x` to `2.x` version of `semantic-ui-react`
 - Added support for `schema.examples` in the material ui theme fixing (https://github.com/rjsf-team/react-jsonschema-form/issues/2368, https://github.com/rjsf-team/react-jsonschema-form/issues/2557)
 
+## @rjsf/bootstrap-4
+- Avoid importing the whole of `react-icons` (https://github.com/rjsf-team/react-jsonschema-form/pull/3046, https://github.com/react-icons/react-icons/issues/154)
+
 ## Dev / docs / playground
 - Fixed missing `playground` import error by adding `source-map-loader`
 - Fixed up the incorrectly formatted `5.x Migration Guide`